+++ package.json
@@ -11,8 +11,6 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.5.2",
     "@types/node": "^16.18.54",
-    "@types/react": "^18.2.22",
-    "@types/react-dom": "^18.2.7",
     "@uiw/react-textarea-code-editor": "^2.1.9",
     "axios": "^1.5.1",
     "bootstrap": "^5.3.2",
@@ -22,8 +20,7 @@
     "react-bootstrap": "^2.8.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.17.0",
-    "react-scripts": "5.0.1",
-    "typescript": "^4.9.5"
+    "react-scripts": "5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -48,5 +45,10 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/react": "^18.2.37",
+    "@types/react-dom": "^18.2.15",
+    "typescript": "^5.2.2"
   }
 }
